[Core] Fix the failing gpu test for the zero-copy tensor
