+++ python/ray/_common/test_utils.py
@@ -282,3 +282,72 @@ def is_named_tuple(cls):
     if not isinstance(f, tuple):
         return False
     return all(type(n) is str for n in f)
+
+
+def assert_tensors_equivalent(obj1, obj2):
+    """
+    Recursively compare objects with special handling for torch.Tensor.
+
+    Tensors are considered equivalent if:
+      - Same dtype and shape
+      - Same device type (e.g., both 'cpu' or both 'cuda'), index ignored
+      - Values are equal (or close for floats)
+    """
+    import torch
+
+    if isinstance(obj1, torch.Tensor) and isinstance(obj2, torch.Tensor):
+        # 1. dtype
+        assert obj1.dtype == obj2.dtype, f"dtype mismatch: {obj1.dtype} vs {obj2.dtype}"
+        # 2. shape
+        assert obj1.shape == obj2.shape, f"shape mismatch: {obj1.shape} vs {obj2.shape}"
+        # 3. device type must match (cpu/cpu or cuda/cuda), ignore index
+        assert (
+            obj1.device.type == obj2.device.type
+        ), f"Device type mismatch: {obj1.device} vs {obj2.device}"
+
+        # 4. Compare values safely on CPU
+        t1_cpu = obj1.cpu()
+        t2_cpu = obj2.cpu()
+        if obj1.dtype.is_floating_point or obj1.dtype.is_complex:
+            assert torch.allclose(
+                t1_cpu, t2_cpu, atol=1e-6, rtol=1e-5
+            ), "Floating-point tensors not close"
+        else:
+            assert torch.equal(t1_cpu, t2_cpu), "Integer/bool tensors not equal"
+        return
+
+    # Type must match
+    if type(obj1) is not type(obj2):
+        raise AssertionError(f"Type mismatch: {type(obj1)} vs {type(obj2)}")
+
+    # Handle namedtuples
+    if is_named_tuple(type(obj1)):
+        assert len(obj1) == len(obj2)
+        for a, b in zip(obj1, obj2):
+            assert_tensors_equivalent(a, b)
+    elif isinstance(obj1, dict):
+        assert obj1.keys() == obj2.keys()
+        for k in obj1:
+            assert_tensors_equivalent(obj1[k], obj2[k])
+    elif isinstance(obj1, (list, tuple)):
+        assert len(obj1) == len(obj2)
+        for a, b in zip(obj1, obj2):
+            assert_tensors_equivalent(a, b)
+    elif hasattr(obj1, "__dict__") and hasattr(obj2, "__dict__"):
+        # Compare user-defined objects by their public attributes
+        keys1 = {
+            k
+            for k in obj1.__dict__.keys()
+            if not k.startswith("_ray_") and k != "_pytype_"
+        }
+        keys2 = {
+            k
+            for k in obj2.__dict__.keys()
+            if not k.startswith("_ray_") and k != "_pytype_"
+        }
+        assert keys1 == keys2, f"Object attribute keys differ: {keys1} vs {keys2}"
+        for k in keys1:
+            assert_tensors_equivalent(obj1.__dict__[k], obj2.__dict__[k])
+    else:
+        # Fallback for primitives: int, float, str, bool, etc.
+        assert obj1 == obj2, f"Non-tensor values differ: {obj1} vs {obj2}"
